requirements.txt
branca==0.7.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
et-xmlfile==1.1.0
Flask==2.2.5
folium==0.16.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==8.0.33
numpy==1.21.6
openpyxl==3.1.3
pandas==1.3.5
protobuf==3.20.3
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
xyzservices==2023.2.0
zipp==3.15.0

python-dotenv~=1.0.1
